+++ packages/web/src/db/models/index.ts
@@ -1,3 +1,4 @@
 export { default as App } from '@/db/models/App';
 export { default as AppUsers } from '@/db/models/AppUsers';
 export { default as User } from '@/db/models/User';
+export { default as Stats } from '@/db/models/Stats';
